data_warehouse/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Vincent <vincent.chery@m4x.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
prefect = "1.3"
pandas = {extras = ["postgresql"], version = "^2.1.3"}
duckdb = "^0.9.2"
black = "^23.11.0"
isort = "^5.12.0"
coverage = "^7.3.2"
pre-commit = "^3.5.0"
flake8 = "^6.1.0"
python-dotenv = "^1.0.0"
pytest = "^7.4.3"
geopandas = "^0.14.1"
simplejson = "^3.19.2"
geoalchemy2 = "^0.14.2"
clickhouse-connect = "^0.6.22"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"